requirements.txt
# Kafka MCP Server Requirements
mcp>=1.0.0
kafka-python>=2.0.2
pydantic>=2.0.0

